+++ src/app/management/dialogs/product-list-contents/product-list-contents-dialog.service.ts
@@ -9,7 +9,7 @@ import { Inject, Injectable, OnDestroy } from '@angular/core';
 import { BehaviorSubject, ReplaySubject, Subscription } from 'rxjs';
 import { finalize, map, tap } from 'rxjs/operators';
 import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-tokens';
-import { IProductListContentsDataApiService } from 'src/app/api/transactional-product-lists.data.api.iservice';
+import { ITransactionalProductListContentsDataApiService } from 'src/app/api/transactional-product-list-contents.data.api.iservice';
 import { DataPage } from 'src/models/DataPage';
 import { Product } from 'src/models/entities/Product';
 import { ProductList } from 'src/models/entities/ProductList';
@@ -33,7 +33,7 @@ export class ProductListContentsDialogService
   list: ProductList | null;
 
   constructor(
-    @Inject(API_SERVICE_INJECTION_TOKENS.dataProductLists) private listApiService: IProductListContentsDataApiService,
+    @Inject(API_SERVICE_INJECTION_TOKENS.dataProductLists) private listApiService: ITransactionalProductListContentsDataApiService,
   ) {
   }