+++ src/app/store/routes/catalog/store-catalog.service.spec.ts
@@ -11,13 +11,13 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { of, throwError } from 'rxjs';
 import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-tokens';
 import { ITransactionalEntityDataApiService } from 'src/app/api/transactional-entity.data-api.iservice';
-import { IProductListContentsDataApiService } from 'src/app/api/transactional-product-lists.data.api.iservice';
+import { ITransactionalProductListContentsDataApiService } from 'src/app/api/transactional-product-list-contents.data.api.iservice';
 import { Product } from 'src/models/entities/Product';
 import { StoreCatalogService } from './store-catalog.service';
 
 describe('StoreCatalogService', () => {
   let service: StoreCatalogService;
-  let mockProductListsApiService: Partial<IProductListContentsDataApiService>;
+  let mockProductListsApiService: Partial<ITransactionalProductListContentsDataApiService>;
   let mockProductsApiService: Partial<ITransactionalEntityDataApiService<Product>>;
   let mockDialogService: Partial<MatDialog>;